Fix min/max map functions
Since these work off of per-cpu maps, it means that when a key
is added the value for the key across all cpus initializes to 0.
Therefore if we loop over all cpus to determine if a value is
greater or less than another we need to discount all 0 values
that weren't set explicitly by the user/prog (e.g. min(0)).

Alternatively, we can simplify min/max by making them use a
hash map, which is what this commit does in addition to using
a less racy value setting (an atomic xchg).

This also preserves the ability to loop over min/max maps,
which was removed in this alternate solution where we kept them
as per-cpu-maps (https://github.com/bpftrace/bpftrace/pull/3288).

Issue: #3286