Add support for named tuple fields
Example:
```
$a = (.name="Jordan", .age=28);
print($a.name); // prints Jordan
```

This still maintains the type as a
tuple instead of a record so this is
fine.

```
$a = (.name="Jordan", .age=28);
$a = ("Viktor", 75);
```

However if there is a field name mismatch
then it's a type error, e.g.,

```
$a = (.name="Jordan", .age=28);
$a = (.person="Viktor", .age=75); // Error
```

When creating a tuple literal all the fields
have to be named or none of them.

Also, named tuples will be printed as records:

```
$a = (.name="Jordan", .age=28);
print($a); // prints { .name = Jordan, .age = 28 }
```

Signed-off-by: Jordan Rome <linux@jordanrome.com>